Net: extract runtime fetch helper
